+++ test_product_queries.py
@@ -1,4 +1,4 @@
-from utils import find_what_user_expects, debug_attribute, resmed_chatbot
+from chatbot import find_what_user_expects, debug_attribute,resmed_chatbot
 from constants import MESSAGE_LOG
 import pandas as pd